Moved the authentication startup procedure into its own class to clean up the main application startup.
