+++ source/gpconnect-appointment-checker/Startup.cs
@@ -37,7 +37,7 @@ namespace gpconnect_appointment_checker
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddHttpContextAccessor();
-            AddAuthenticationServices(services); 
+            services.ConfigureAuthenticationServices(Configuration);
             services.ConfigureApplicationServices(Configuration, WebHostEnvironment);
             services.ConfigureLoggingServices(Configuration);
             MappingExtensions.ConfigureMappingServices();
@@ -48,59 +48,5 @@ namespace gpconnect_appointment_checker
             _applicationService = applicationService;
             app.ConfigureApplicationBuilderServices(env);
         }
-
-        public void AddAuthenticationServices(IServiceCollection services)
-        {
-            services.AddAuthentication(options =>
-            {
-                options.DefaultScheme = CookieAuthenticationDefaults.AuthenticationScheme;
-                options.DefaultChallengeScheme = Configuration.GetSection("SingleSignOn:challenge_scheme").GetConfigurationString(throwExceptionIfEmpty: true);
-                options.DefaultSignOutScheme = Configuration.GetSection("SingleSignOn:challenge_scheme").GetConfigurationString(throwExceptionIfEmpty: true);
-            }).AddCookie(s => {
-                s.ExpireTimeSpan = TimeSpan.FromMinutes(30);
-                s.SlidingExpiration = true;
-                })
-            .AddOpenIdConnect(
-                Configuration.GetSection("SingleSignOn:challenge_scheme").GetConfigurationString(throwExceptionIfEmpty: true),
-                displayName: Configuration.GetSection("SingleSignOn:challenge_scheme").GetConfigurationString(throwExceptionIfEmpty: true),
-                options =>
-                {
-                    options.RequireHttpsMetadata = true;
-                    options.ResponseMode = OpenIdConnectResponseMode.FormPost;
-                    options.Authority = Configuration.GetSection("SingleSignOn:auth_endpoint").GetConfigurationString(throwExceptionIfEmpty: true);
-                    options.MetadataAddress = Configuration.GetSection("SingleSignOn:metadata_endpoint").GetConfigurationString(throwExceptionIfEmpty: true);
-                    options.ClientId = Configuration.GetSection("SingleSignOn:client_id").GetConfigurationString(throwExceptionIfEmpty: true);
-                    options.ClientSecret = Configuration.GetSection("SingleSignOn:client_secret").GetConfigurationString(throwExceptionIfEmpty: true);
-                    options.CallbackPath = Configuration.GetSection("SingleSignOn:callback_path").GetConfigurationString(throwExceptionIfEmpty: true);
-                    //options.RemoteSignOutPath = "/Public/Index";
-                    //options.SignedOutRedirectUri = "/Public/Index";
-                    options.ResponseType = OpenIdConnectResponseType.IdToken;
-                    options.Scope.Add("email");
-                    options.Scope.Add("profile");
-                    options.Scope.Add("openid");
-                    options.GetClaimsFromUserInfoEndpoint = true;
-                    options.Events = new OpenIdConnectEvents
-                    {
-                        OnTokenValidated = context =>
-                        {
-                            var ldapTokenService = new LdapTokenService(services.BuildServiceProvider());
-                            var tokenValidation = ldapTokenService.ExecutionTokenValidation(context);
-                            return tokenValidation;
-                        },
-                        OnAuthenticationFailed = context =>
-                        {
-                            if (context.Exception == null)
-                            {
-                                context.Response.Redirect("/AccessDenied");
-                            }
-                            context.Response.Redirect("/Error");
-                            context.HandleResponse();
-                            return Task.CompletedTask;
-                        }
-                    };
-                });
-        }
-
-        
     }
 }
